+++ internal/gengapic/generator.go
@@ -36,10 +36,9 @@ import (
 	"google.golang.org/protobuf/types/pluginpb"
 )
 
-var enableNewAuthLibrary = map[string]bool{
-	"cloudkms.googleapis.com":      true,
-	"secretmanager.googleapis.com": true,
-	"showcase.googleapis.com":      true,
+// keyed by proto package name, e.g. "google.cloud.foo.v1".
+var enableWrapperTypesForPageSize = map[string]bool{
+	"google.cloud.bigquery.v2": true,
 }
 
 type generator struct {
@@ -203,12 +202,19 @@ func (g *generator) printf(s string, a ...interface{}) {
 
 // TODO(chrisdsmith): Add generator_test.go with TestCommit
 
+func (g *generator) commit(fileName, pkgName string) int {
+	return g.commitWithBuildTag(fileName, pkgName, "")
+}
+
 // commit adds header, etc to current pt and returns the line length of the
 // final file output.
-func (g *generator) commit(fileName, pkgName string) int {
+func (g *generator) commitWithBuildTag(fileName, pkgName, buildTag string) int {
 	var header strings.Builder
 	fmt.Fprintf(&header, license.Apache, time.Now().Year())
 	header.WriteString(g.headerComments.String() + "\n")
+	if buildTag != "" {
+		fmt.Fprintf(&header, "//go:build %s\n\n", buildTag)
+	}
 	fmt.Fprintf(&header, "package %s\n\n", pkgName)
 
 	var imps []pbinfo.ImportSpec
@@ -343,4 +349,20 @@ func (g *generator) autoPopulatedFields(_ string, m *descriptorpb.MethodDescript
 		}
 	}
 	return validated
+}
+
+// getServiceNameOverride checks to see if the service has a defined service name override.
+func (g *generator) getServiceNameOverride(s *descriptorpb.ServiceDescriptorProto) string {
+	ls := g.serviceConfig.GetPublishing().GetLibrarySettings()
+	if len(ls) == 0 {
+		return ""
+	}
+
+	renamedServices := ls[0].GetGoSettings().GetRenamedServices()
+
+	if v, ok := renamedServices[s.GetName()]; ok {
+		return v
+	}
+
+	return ""
 }