+++ internal/gengapic/genrest.go
@@ -91,6 +91,8 @@ func (g *generator) restClientInit(serv *descriptorpb.ServiceDescriptorProto, se
 	p("")
 	p("  // Points back to the CallOptions field of the containing %sClient", servName)
 	p("  CallOptions **%sCallOptions", servName)
+	p("")
+	p("  logger *slog.Logger")
 	p("}")
 	p("")
 	g.restClientUtilities(serv, servName, hasRPCForLRO)
@@ -98,6 +100,7 @@ func (g *generator) restClientInit(serv *descriptorpb.ServiceDescriptorProto, se
 	g.imports[pbinfo.ImportSpec{Path: "net/http"}] = true
 	g.imports[pbinfo.ImportSpec{Name: "httptransport", Path: "google.golang.org/api/transport/http"}] = true
 	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/api/option/internaloption"}] = true
+	g.imports[pbinfo.ImportSpec{Path: "log/slog"}] = true
 }
 
 func (g *generator) genRESTMethods(serv *descriptorpb.ServiceDescriptorProto, servName string) error {
@@ -137,18 +140,22 @@ func (g *generator) restClientOptions(serv *descriptorpb.ServiceDescriptorProto,
 	p("    internaloption.WithDefaultUniverseDomain(%q),", googleDefaultUniverse)
 	p("    internaloption.WithDefaultAudience(%q),", generateDefaultAudience(host))
 	p("    internaloption.WithDefaultScopes(DefaultAuthScopes()...),")
-	if _, ok := enableNewAuthLibrary[g.serviceConfig.GetName()]; ok {
-		p("internaloption.EnableNewAuthLibrary(),")
-	}
+	p("    internaloption.EnableNewAuthLibrary(),")
 	p("  }")
 	p("}")
 }
 
 func (g *generator) restClientUtilities(serv *descriptorpb.ServiceDescriptorProto, servName string, hasRPCForLRO bool) {
 	p := g.printf
-	lowcaseServName := lowcaseRestClientName(servName)
-	clientName := camelToSnake(serv.GetName())
+
+	clientName := serv.GetName()
+	if override := g.getServiceNameOverride(serv); override != "" {
+		clientName = override
+	}
+	clientName = camelToSnake(clientName)
 	clientName = strings.Replace(clientName, "_", " ", -1)
+	lowcaseServName := lowcaseRestClientName(servName)
+
 	opServ, hasCustomOp := g.customOpServices[serv]
 
 	p("// New%sRESTClient creates a new %s rest client.", servName, clientName)
@@ -165,6 +172,7 @@ func (g *generator) restClientUtilities(serv *descriptorpb.ServiceDescriptorProt
 	p("        endpoint: endpoint,")
 	p("        httpClient: httpClient,")
 	p("        CallOptions: &callOpts,")
+	p("        logger: internaloption.GetLogger(opts),")
 	p("    }")
 	p("    c.setGoogleClientInfo()")
 	p("")
@@ -417,7 +425,7 @@ func (g *generator) generateQueryString(m *descriptorpb.MethodDescriptorProto) {
 		// Handle well known protobuf types with special JSON encodings.
 		if strContains(wellKnownTypeNames, field.GetTypeName()) {
 			b := strings.Builder{}
-			b.WriteString(fmt.Sprintf("%s, err := protojson.Marshal(req%s)\n", field.GetJsonName(), accessor))
+			b.WriteString(fmt.Sprintf("field, err := protojson.Marshal(req%s)\n", accessor))
 			b.WriteString("if err != nil {\n")
 			if m.GetOutputType() == emptyType {
 				b.WriteString("  return err\n")
@@ -429,9 +437,9 @@ func (g *generator) generateQueryString(m *descriptorpb.MethodDescriptorProto) {
 			b.WriteString("}\n")
 			// Only some of the well known types will be encoded as strings, remove the wrapping quotations for those.
 			if strContains(wellKnownStringTypes, field.GetTypeName()) {
-				b.WriteString(fmt.Sprintf("params.Add(%q, string(%s[1:len(%[2]s)-1]))", key, field.GetJsonName()))
+				b.WriteString(fmt.Sprintf("params.Add(%q, string(field[1:len(field)-1]))", key))
 			} else {
-				b.WriteString(fmt.Sprintf("params.Add(%q, string(%s))", key, field.GetJsonName()))
+				b.WriteString(fmt.Sprintf("params.Add(%q, string(field))", key))
 			}
 			paramAdd = b.String()
 		} else {
@@ -619,7 +627,7 @@ func (g *generator) serverStreamRESTCall(servName string, s *descriptorpb.Servic
 	// rest-client method
 	p("func (c *%s) %s(ctx context.Context, req *%s.%s, opts ...gax.CallOption) (%s.%s_%sClient, error) {",
 		lowcaseServName, m.GetName(), inSpec.Name, inType.GetName(), servSpec.Name, s.GetName(), m.GetName())
-	body := "nil"
+	body, logBody := "nil", "nil"
 	verb := strings.ToUpper(info.verb)
 
 	// Marshal body for HTTP methods that take a body.
@@ -640,6 +648,7 @@ func (g *generator) serverStreamRESTCall(servName string, s *descriptorpb.Servic
 		p("")
 
 		body = "bytes.NewReader(jsonReq)"
+		logBody = "jsonReq"
 		g.imports[pbinfo.ImportSpec{Path: "bytes"}] = true
 		g.imports[pbinfo.ImportSpec{Path: "google.golang.org/protobuf/encoding/protojson"}] = true
 	}
@@ -660,13 +669,9 @@ func (g *generator) serverStreamRESTCall(servName string, s *descriptorpb.Servic
 	p("  httpReq = httpReq.WithContext(ctx)")
 	p("  httpReq.Header = headers")
 	p("")
-	p("  httpRsp, err := c.httpClient.Do(httpReq)")
+	p("  httpRsp, err := executeStreamingHTTPRequest(ctx, c.httpClient, httpReq, c.logger, %s, %q)", logBody, m.GetName())
 	p("  if err != nil{")
 	p("   return err")
-	p("  }")
-	p("")
-	p("  if err = googleapi.CheckResponse(httpRsp); err != nil {")
-	p("    return err")
 	p("  }")
 	p("")
 	p("  streamClient = &%s{", streamClient)
@@ -680,8 +685,6 @@ func (g *generator) serverStreamRESTCall(servName string, s *descriptorpb.Servic
 	p("return streamClient, e")
 	p("}")
 	p("")
-
-	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/api/googleapi"}] = true
 
 	// server-stream wrapper client
 	p("// %s is the stream client used to consume the server stream created by", streamClient)
@@ -791,35 +794,28 @@ func (g *generator) pagingRESTCall(servName string, m *descriptorpb.MethodDescri
 
 	verb := strings.ToUpper(info.verb)
 
-	max := "math.MaxInt32"
 	g.imports[pbinfo.ImportSpec{Path: "math"}] = true
-	psTyp := pbinfo.GoTypeForPrim[pageSize.GetType()]
-	ps := fmt.Sprintf("%s(pageSize)", psTyp)
-	if isOptional(inType, pageSize.GetName()) {
-		max = fmt.Sprintf("proto.%s(%s)", upperFirst(psTyp), max)
-		ps = fmt.Sprintf("proto.%s(%s)", upperFirst(psTyp), ps)
-	}
 	tok := "pageToken"
 	if isOptional(inType, "page_token") {
 		tok = fmt.Sprintf("proto.String(%s)", tok)
 	}
 
-	pageSizeFieldName := snakeToCamel(pageSize.GetName())
 	p("func (c *%s) %s(ctx context.Context, req *%s.%s, opts ...gax.CallOption) *%s {",
 		lowcaseServName, m.GetName(), inSpec.Name, inType.GetName(), pt.iterTypeName)
 	p("it := &%s{}", pt.iterTypeName)
 	p("req = proto.Clone(req).(*%s.%s)", inSpec.Name, inType.GetName())
 
-	maybeReqBytes := "nil"
+	maybeReqBytes, logBody := "nil", "nil"
 	if info.body != "" {
 		g.protoJSONMarshaler()
 		maybeReqBytes = "bytes.NewReader(jsonReq)"
+		logBody = "jsonReq"
 		g.imports[pbinfo.ImportSpec{Path: "bytes"}] = true
 	}
 
 	p("unm := protojson.UnmarshalOptions{AllowPartial: true, DiscardUnknown: true}")
 	p("it.InternalFetch = func(pageSize int, pageToken string) ([]%s, string, error) {", pt.elemTypeName)
-	g.internalFetchSetup(outType, outSpec, tok, pageSizeFieldName, max, ps)
+	g.internalFetchSetup(outType, outSpec, pageSize, tok)
 
 	if info.body != "" {
 		p("  jsonReq, err := m.Marshal(req)")
@@ -845,21 +841,10 @@ func (g *generator) pagingRESTCall(servName string, m *descriptorpb.MethodDescri
 	// TODO: Should this http.Request use WithContext?
 	p("    httpReq.Header = headers")
 	p("")
-	p("    httpRsp, err := c.httpClient.Do(httpReq)")
+	p("    buf, err := executeHTTPRequest(ctx, c.httpClient, httpReq, c.logger, %s, %q)", logBody, m.GetName())
 	p("    if err != nil{")
 	p(`     return err`)
 	p("    }")
-	p("    defer httpRsp.Body.Close()")
-	p("")
-	p("    if err = googleapi.CheckResponse(httpRsp); err != nil {")
-	p(`      return err`)
-	p("    }")
-	p("")
-	p("    buf, err := io.ReadAll(httpRsp.Body)")
-	p("    if err != nil {")
-	p(`      return err`)
-	p("    }")
-	p("")
 	p("    if err := unm.Unmarshal(buf, resp); err != nil {")
 	p("      return err")
 	p("    }")
@@ -874,14 +859,12 @@ func (g *generator) pagingRESTCall(servName string, m *descriptorpb.MethodDescri
 	p("  return %s, resp.GetNextPageToken(), nil", elems)
 	p("}")
 	p("")
-	g.makeFetchAndIterUpdate(pageSizeFieldName)
+	g.makeFetchAndIterUpdate(pageSize)
 	p("}")
 
 	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/api/iterator"}] = true
-	g.imports[pbinfo.ImportSpec{Path: "io"}] = true
 	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/protobuf/proto"}] = true
 	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/protobuf/encoding/protojson"}] = true
-	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/api/googleapi"}] = true
 	g.imports[inSpec] = true
 	g.imports[outSpec] = true
 
@@ -926,7 +909,7 @@ func (g *generator) lroRESTCall(servName string, m *descriptorpb.MethodDescripto
 	// TODO(noahdietz): handle deadlines?
 	// TODO(noahdietz): handle calloptions
 
-	body := "nil"
+	body, logBody := "nil", "nil"
 	verb := strings.ToUpper(info.verb)
 
 	// Marshal body for HTTP methods that take a body.
@@ -947,6 +930,7 @@ func (g *generator) lroRESTCall(servName string, m *descriptorpb.MethodDescripto
 		p("")
 
 		body = "bytes.NewReader(jsonReq)"
+		logBody = "jsonReq"
 		g.imports[pbinfo.ImportSpec{Path: "bytes"}] = true
 	}
 
@@ -967,21 +951,10 @@ func (g *generator) lroRESTCall(servName string, m *descriptorpb.MethodDescripto
 	p("  httpReq = httpReq.WithContext(ctx)")
 	p("  httpReq.Header = headers")
 	p("")
-	p("  httpRsp, err := c.httpClient.Do(httpReq)")
+	p("  buf, err := executeHTTPRequest(ctx, c.httpClient, httpReq, c.logger, %s, %q)", logBody, m.GetName())
 	p("  if err != nil{")
 	p("   return err")
 	p("  }")
-	p("  defer httpRsp.Body.Close()")
-	p("")
-	p("  if err = googleapi.CheckResponse(httpRsp); err != nil {")
-	p("    return err")
-	p("  }")
-	p("")
-	p("  buf, err := io.ReadAll(httpRsp.Body)")
-	p("  if err != nil {")
-	p("    return err")
-	p("  }")
-	p("")
 	p("  if err := unm.Unmarshal(buf, resp); err != nil {")
 	p("    return err")
 	p("  }")
@@ -1002,9 +975,7 @@ func (g *generator) lroRESTCall(servName string, m *descriptorpb.MethodDescripto
 	p("")
 
 	g.imports[pbinfo.ImportSpec{Path: "fmt"}] = true
-	g.imports[pbinfo.ImportSpec{Path: "io"}] = true
 	g.imports[pbinfo.ImportSpec{Path: "cloud.google.com/go/longrunning"}] = true
-	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/api/googleapi"}] = true
 	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/protobuf/encoding/protojson"}] = true
 
 	return nil
@@ -1033,7 +1004,7 @@ func (g *generator) emptyUnaryRESTCall(servName string, m *descriptorpb.MethodDe
 	// TODO(dovs): handle deadlines
 	// TODO(dovs): handle call options
 
-	body := "nil"
+	body, logBody := "nil", "nil"
 	verb := strings.ToUpper(info.verb)
 
 	// Marshal body for HTTP methods that take a body.
@@ -1054,6 +1025,7 @@ func (g *generator) emptyUnaryRESTCall(servName string, m *descriptorpb.MethodDe
 		p("}")
 		p("")
 		body = "bytes.NewReader(jsonReq)"
+		logBody = "jsonReq"
 		g.imports[pbinfo.ImportSpec{Path: "bytes"}] = true
 		g.imports[pbinfo.ImportSpec{Path: "google.golang.org/protobuf/encoding/protojson"}] = true
 	}
@@ -1073,19 +1045,11 @@ func (g *generator) emptyUnaryRESTCall(servName string, m *descriptorpb.MethodDe
 	p("  httpReq = httpReq.WithContext(ctx)")
 	p("  httpReq.Header = headers")
 	p("")
-	p("  httpRsp, err := c.httpClient.Do(httpReq)")
-	p("  if err != nil{")
-	p("   return err")
-	p("  }")
-	p("  defer httpRsp.Body.Close()")
-	p("")
-	p("  // Returns nil if there is no error, otherwise wraps")
-	p("  // the response code and body into a non-nil error")
-	p("  return googleapi.CheckResponse(httpRsp)")
+	p("  _, err = executeHTTPRequest(ctx, c.httpClient, httpReq, c.logger, %s, %q)", logBody, m.GetName())
+	p("  return err")
 	p("  }, opts...)")
 	p("}")
 
-	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/api/googleapi"}] = true
 	g.imports[inSpec] = true
 	return nil
 }
@@ -1127,7 +1091,7 @@ func (g *generator) unaryRESTCall(servName string, m *descriptorpb.MethodDescrip
 	// TODO(dovs): handle deadlines?
 	// TODO(dovs): handle calloptions
 
-	body := "nil"
+	body, logBody := "nil", "nil"
 	verb := strings.ToUpper(info.verb)
 
 	// Marshal body for HTTP methods that take a body.
@@ -1149,6 +1113,7 @@ func (g *generator) unaryRESTCall(servName string, m *descriptorpb.MethodDescrip
 		p("")
 
 		body = "bytes.NewReader(jsonReq)"
+		logBody = "jsonReq"
 		g.imports[pbinfo.ImportSpec{Path: "bytes"}] = true
 		g.imports[pbinfo.ImportSpec{Path: "google.golang.org/protobuf/encoding/protojson"}] = true
 
@@ -1176,19 +1141,13 @@ func (g *generator) unaryRESTCall(servName string, m *descriptorpb.MethodDescrip
 	p("  httpReq = httpReq.WithContext(ctx)")
 	p("  httpReq.Header = headers")
 	p("")
-	p("  httpRsp, err := c.httpClient.Do(httpReq)")
+	if isHTTPBodyMessage {
+		p("  buf, httpRsp, err := executeHTTPRequestWithResponse(ctx, c.httpClient, httpReq, c.logger, %s, %q)", logBody, m.GetName())
+	} else {
+		p("  buf, err := executeHTTPRequest(ctx, c.httpClient, httpReq, c.logger, %s, %q)", logBody, m.GetName())
+	}
 	p("  if err != nil{")
 	p("   return err")
-	p("  }")
-	p("  defer httpRsp.Body.Close()")
-	p("")
-	p("  if err = googleapi.CheckResponse(httpRsp); err != nil {")
-	p("    return err")
-	p("  }")
-	p("")
-	p("  buf, err := io.ReadAll(httpRsp.Body)")
-	p("  if err != nil {")
-	p("    return err")
 	p("  }")
 	p("")
 	if isHTTPBodyMessage {
@@ -1217,8 +1176,6 @@ func (g *generator) unaryRESTCall(servName string, m *descriptorpb.MethodDescrip
 	p(ret)
 	p("}")
 
-	g.imports[pbinfo.ImportSpec{Path: "io"}] = true
-	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/api/googleapi"}] = true
 	g.imports[inSpec] = true
 	g.imports[outSpec] = true
 	return nil